# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "TiOS API",
        "description": "Welcolme to TiOS Platfor API documentation.",
        "termsOfService": "",
        "contact": {
            "email": "luis@tiosplatform.com"
        },
        "license": {
            "name": "Private License",
            "url": "URL to the license"
        },
        "version": "1.0.0"
    },
    "host": "127.0.0.1:5000/api",
    "basePath": "/",
    "schemes": [
        "http",
        "https"
    ],
    "paths": {
...
        "/event/create": {
            "post": {
                "tags": [
                    "Events"
                ],
                "summary": "Create new event entry",
                "operationId": "createEventDB",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "JSON Payload",
                        "required": true,
                        "schema": {
                            "properties": {
                                "event_id": {
                                    "description": "event id",
                                    "type": "integer",
                                    "example": "1512581640"
                                },
                                "title": {
                                    "type": "string",
                                    "example": "Send email"
                                },
                                "action": {
                                    "type": "string",
                                    "example": "[{'action':'send-email','recipients':'[6]','message':'custom message','triggered':0}]"
                                },
                                "valueFrom": {
                                    "type": "string",
                                    "example": "1512581640"
                                },
                                "color": {
                                    "type": "string",
                                    "example": "#fac5a5"
                                },
                                "organization_id": {
                                    "type": "integer",
                                    "example": "1"
                                },
                                "project_id": {
                                    "type": "integer",
                                    "example": "1"
                                },
                                "datasource_id": {
                                    "type": "integer",
                                    "example": "1"
                                },
                                "datapoint_id": {
                                    "type": "integer",
                                    "example": "1"
                                },
                                "active": {
                                    "type": "integer",
                                    "example": "1"
                                },
                                "end": {
                                    "type": "string",
                                    "example": "end"
                                }
                            },
                            "type": "object"
                        },
                        "type": "json",
                        "format": "application/json"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/events": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "Get list of events",
                "description": "Returns list of events",
                "operationId": "getEventsDB",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/event/{id}": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "Get event information",
                "description": "Returns event data",
                "operationId": "getEventDB",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "event id",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/events/organization/{organization_id}": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "Get event information By Organization",
                "description": "Returns event data by organization",
                "operationId": "getEventsDBByOrgId",
                "parameters": [
                    {
                        "name": "organization_id",
                        "in": "path",
                        "description": "organization id",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/events/actions": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "Get list of events actions",
                "description": "Returns list of events actions",
                "operationId": "getEventsDBActions",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
...
    },
...
    "securityDefinitions": {
        "passport": {
            "type": "oauth2",
            "flow": "password",
            "tokenUrl": "/oauth/token",
            "scopes": {}
        }
    },
...
}